+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+
+## 0.0.2
+* Added ``example.dart`` for how to use ``mathx`` package
+
 ## 0.0.1
--- Initial release of the mathx package.
+* Initial release of the ``mathx`` package.